chore: fixes 2
