+++ src/entities/app.rs
@@ -3,11 +3,8 @@ use std::collections::BTreeMap;
 use iced::window::Id;
 use tray_icon::TrayIcon;
 
-use crate::entities::{
-    capture::CaptureEvent,
-    config::{Config, ConfigEvent},
-    window::WindowType,
-};
+pub use crate::entities::events::AppEvent;
+use crate::entities::{config::Config, window::WindowType};
 
 pub struct App {
     #[expect(dead_code)]
@@ -17,32 +14,11 @@ pub struct App {
 }
 
 impl App {
-    pub(crate) fn new_internal(
-        tray_icon: TrayIcon,
-        config: Config,
-        windows: BTreeMap<Id, WindowType>,
-    ) -> Self {
+    pub fn new(tray_icon: TrayIcon, config: Config) -> Self {
         Self {
             tray_icon,
             config,
-            windows,
+            windows: BTreeMap::new(),
         }
     }
-}
-
-#[derive(Debug, Clone)]
-pub enum AppEvent {
-    OpenConfigureWindow,
-    OpenDirectory,
-    UpdateDirectory(Id),
-    GetScaleFactor(Id, f32),
-    OpenCaptureWindow,
-    Undo,
-    Done,
-    Cancel,
-    RequestClose(Id),
-    WindowClosed(Id),
-    ExitApp,
-    Config(Id, ConfigEvent),
-    Capture(Id, CaptureEvent),
 }
