+++ src/entities/mod.rs
@@ -1,6 +1,7 @@
 pub mod app;
 pub mod capture;
 pub mod config;
+mod events;
 pub mod style;
 pub mod theme;
 pub mod window;
